# packages/backend/src/assets/openapi.yaml
openapi: 3.1.0
info:
  title: chickenScratch API — Biometric Account Recovery
  version: 1.0.0
  description: |
    chickenScratch is a drop-in biometric account recovery service. When your
    users forget their password — or which email they used to sign up — they
    sign their name to prove it's really them, instead of going through an
    email reset link or SMS OTP. Integrate via the JavaScript SDK or the REST
    API below.

    The same enrollment/verify primitives also support step-up auth for
    sensitive actions (wire approvals, settings changes, high-value transfers).
    Recovery is the primary use case we're building for; everything below
    applies to both.

    ## Authentication

    All `/api/v1/*` endpoints accept either a tenant API key (server-to-server)
    or a short-lived SDK token (browser use).

    - API key: `X-API-Key: cs_live_xxxxxxxxxxxxxxxxxxxxxxxx`
    - SDK token: `Authorization: Bearer cs_sdk_xxxxxxxxxxxxxxxxxxxx`

    API keys are issued once at tenant creation. SDK tokens are minted by your
    backend via `POST /api/v1/sdk-token` and expire in 15 minutes — use them
    to keep your API key off the browser.

    ## Consent requirement

    chickenScratch collects biometric identifiers. Before any enrollment data
    is accepted, the end user must give **explicit consent** via
    `POST /api/v1/consent`. Enrollment calls return `403 Forbidden` until
    consent is recorded. This is required for BIPA (Illinois), Texas CUBI,
...

...

    ```
...

...

...

...

...

...

...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
...
    ShapeType:
      type: string
      enum: [circle, square, triangle, house, smiley]
      description: The shape the user was asked to draw
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Explicit consent required before biometric enrollment.

# ─── Security (applied globally to /api/v1/* routes) ─────────────────────────
security: []   # overridden per-operation
...
paths:
...
  /api/v1/consent:
    post:
      summary: Record consent
      operationId: recordConsent
      tags: [Consent]
      security:
        - ApiKeyAuth: []
      description: |
        Record explicit biometric data consent for a user. Must be called
        before any enrollment. Calling this again for a user who has already
        consented refreshes the timestamp (idempotent).

        The `policyVersion` field defaults to the current version (`1.0`) if
        omitted. If the policy version changes in future, users will need to
        re-consent before enrollment resumes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [externalUserId]
              properties:
                externalUserId:
                  type: string
                  maxLength: 255
                  description: Your system's identifier for this user
                  example: user_abc123
                policyVersion:
                  type: string
                  default: '1.0'
                  description: Policy version the user is consenting to
            example:
              externalUserId: user_abc123
              policyVersion: '1.0'
      responses:
        '200':
          description: Consent recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  externalUserId:
                    type: string
                  policyVersion:
                    type: string
                  consentedAt:
                    type: string
                    format: date-time
                  message:
                    type: string
                    example: Consent recorded.
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/consent/{externalUserId}:
    get:
      summary: Check consent status
      operationId: getConsentStatus
      tags: [Consent]
      security:
        - ApiKeyAuth: []
      description: |
        Returns the user's current consent status. Use `hasConsented` to gate
        your enrollment UI — only show it when this is `true`.
      parameters:
        - name: externalUserId
          in: path
          required: true
          schema:
            type: string
          example: user_abc123
      responses:
        '200':
          description: Consent status
          content:
            application/json:
              schema:
                type: object
                properties:
                  externalUserId:
                    type: string
                  hasConsented:
                    type: boolean
                    description: True if the user has active consent for the current policy version
                  policyVersion:
                    type: string
                    nullable: true
                  consentedAt:
                    type: string
                    format: date-time
                    nullable: true
                  isCurrentVersion:
                    type: boolean
                    description: False if the user consented to an older policy version
                  currentPolicyVersion:
                    type: string
                    example: '1.0'
              example:
                externalUserId: user_abc123
                hasConsented: true
                policyVersion: '1.0'
                consentedAt: '2026-04-14T12:00:00.000Z'
                isCurrentVersion: true
                currentPolicyVersion: '1.0'

    delete:
      summary: Withdraw consent
      operationId: withdrawConsent
      tags: [Consent]
      security:
        - ApiKeyAuth: []
      description: |
        Withdraw the user's consent and **immediately delete all their
        biometric data** (enrollment samples, baselines, auth attempts).
        Consent records are preserved for 7 years as required by BIPA/GDPR.

        After withdrawal, any further enrollment calls will return `403` until
        the user re-consents.
      parameters:
        - name: externalUserId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Consent withdrawn and biometric data deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  externalUserId:
                    type: string
                  message:
                    type: string
                  deletionSummary:
                    type: object
                    description: Count of records deleted per table
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Enrollment ─────────────────────────────────────────────────────────────
  /api/v1/enroll:
    post:
      summary: Submit a signature sample
      operationId: enrollSignature
      tags: [Enrollment]
      security:
        - ApiKeyAuth: []
      description: |
        Submit one signature drawing. Call this **3 times** with the user
        drawing their signature naturally each time. After the 3rd sample,
        a biometric baseline is computed automatically.

        Returns `enrolled: false` with `samplesRemaining > 0` until all
        3 samples are collected. Shape enrollment must follow before the
        user is fully enrolled.

        **Requires active consent** — returns `403` if the user has not
        consented or has withdrawn consent.

        **Quality gates** — returns `400` if the sample is too short
        (< 200ms), has too few data points (< 20), or covers too small an
        area (< 30px in both dimensions). Prompt the user to redraw.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [externalUserId, signatureData]
              properties:
                externalUserId:
                  type: string
                  example: user_abc123
                signatureData:
                  $ref: '#/components/schemas/RawSignatureData'
                skipRecentVerify:
                  type: boolean
                  default: false
                  description: |
                    Opt out of the add-device recent-verify gate. Only pass
                    `true` if you have authenticated the user through other
                    means (password + MFA, etc.). Ignored for users with no
                    existing baselines (initial enrollment).
      responses:
        '200':
          description: Sample recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  externalUserId:
                    type: string
                  sampleNumber:
                    type: integer
                    description: Which sample this was (1, 2, or 3)
                  samplesRemaining:
                    type: integer
                    description: How many more signature samples are needed
                  enrolled:
                    type: boolean
                    description: True only after all signatures AND shapes are enrolled
                  message:
                    type: string
                  deviceClass:
                    type: string
                    enum: [mobile, desktop]
                    description: The device class detected from this sample.
                  enrolledClasses:
                    type: array
                    items:
                      type: string
                      enum: [mobile, desktop]
                    description: All classes this user has baselines for.
              example:
                success: true
                externalUserId: user_abc123
                sampleNumber: 1
                samplesRemaining: 2
                enrolled: false
                message: Sample 1 of 3 recorded.
                deviceClass: mobile
        '400':
          description: Sample rejected (quality gate failed, validation error, or already enrolled on this device class)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: |
            Two distinct cases:
            1. `errorCode: RECENT_VERIFY_REQUIRED` — user has existing baselines on other classes, needs to verify first before adding this class. Their last successful verify is older than the 10-minute window. Bypass with `skipRecentVerify: true` if you've authenticated them another way.
            2. No errorCode — user has not recorded consent via `POST /api/v1/consent`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                  errorCode:
                    type: string
                    enum: [RECENT_VERIFY_REQUIRED]
                  enrolledClasses:
                    type: array
                    items:
                      type: string
                      enum: [mobile, desktop]
              examples:
                recentVerify:
                  summary: Add-device gate
                  value:
                    success: false
                    message: Adding a new device class requires a recent successful verification on an already-enrolled device.
                    errorCode: RECENT_VERIFY_REQUIRED
                    enrolledClasses: [mobile]
                consent:
                  summary: Consent missing
                  value:
                    success: false
                    error: Explicit consent required before biometric enrollment. Call POST /api/v1/consent first.
  /api/v1/enroll/shape:
    post:
      summary: Submit a shape drawing
      operationId: enrollShape
      tags: [Enrollment]
      security:
        - ApiKeyAuth: []
      description: |
        Submit one shape drawing for enrollment. Call once for each of the
        5 shapes: `circle`, `square`, `triangle`, `house`, `smiley`.

        Signature enrollment (3 samples) must be completed before shapes
        are accepted.

        After all 5 shapes are enrolled, `enrolled` becomes `true` on the
        status endpoint and the user can proceed to verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [externalUserId, shapeType, signatureData]
              properties:
                externalUserId:
                  type: string
                  example: user_abc123
                shapeType:
                  $ref: '#/components/schemas/ShapeType'
                signatureData:
                  $ref: '#/components/schemas/RawSignatureData'
            example:
              externalUserId: user_abc123
              shapeType: circle
      responses:
        '200':
          description: Shape enrolled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  externalUserId:
                    type: string
                  message:
                    type: string
              example:
                success: true
                externalUserId: user_abc123
                message: Shape 'circle' enrolled successfully.
        '400':
          description: Quality gate failed or signature not yet enrolled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Consent required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/enroll/{externalUserId}/status:
    get:
      summary: Get enrollment status
      operationId: getEnrollmentStatus
      tags: [Enrollment]
      security:
        - ApiKeyAuth: []
      description: |
        Check how far through enrollment a user is. Poll this after each
        enrollment step, or use it to resume an interrupted flow.

        `enrolled: true` means both signature (3 samples) and all 5 shapes
        are complete — the user is ready to verify.
      parameters:
        - name: externalUserId
          in: path
          required: true
          schema:
            type: string
          example: user_abc123
      responses:
        '200':
          description: Enrollment status
          content:
            application/json:
              schema:
                type: object
                properties:
                  externalUserId:
                    type: string
                  enrolled:
                    type: boolean
                  samplesCollected:
                    type: integer
                  samplesRequired:
                    type: integer
                    example: 3
                  shapesEnrolled:
                    type: array
                    items:
                      type: string
                    example: [circle, square]
                  shapesRequired:
                    type: array
                    items:
                      type: string
                    example: [circle, square, triangle, house, smiley]
              example:
                externalUserId: user_abc123
                enrolled: false
                samplesCollected: 3
                samplesRequired: 3
                shapesEnrolled: [circle, square, triangle]
                shapesRequired: [circle, square, triangle, house, smiley]

  # ── Verification ───────────────────────────────────────────────────────────
  /api/v1/challenge:
    post:
      summary: Create a verification challenge
      operationId: createChallenge
      tags: [Verification]
      security:
        - ApiKeyAuth: []
      description: |
        Create a one-time challenge before collecting verification drawings.
        Returns a `challengeId` and a randomised `shapeOrder` that the user
        must draw in exactly that sequence.

        Challenges expire after **5 minutes**. Generate a new one for each
        verification attempt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [externalUserId]
              properties:
                externalUserId:
                  type: string
                  example: user_abc123
      responses:
        '200':
          description: Challenge created
          content:
            application/json:
              schema:
                type: object
                properties:
                  challengeId:
                    type: string
                    format: uuid
                    description: Pass this back in POST /api/v1/verify
                  shapeOrder:
                    type: array
                    items:
                      $ref: '#/components/schemas/ShapeType'
                    description: Draw shapes in this exact order
                  expiresAt:
                    type: string
                    format: date-time
              example:
                challengeId: 550e8400-e29b-41d4-a716-446655440000
                shapeOrder: [triangle, smiley, circle, house, square]
                expiresAt: '2026-04-14T12:05:00.000Z'
        '404':
          description: User not found or not enrolled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/tenants/{id}:
    get:
      summary: Get tenant details
      operationId: getTenant
      tags: [Admin]
      security:
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant details including enrolled users
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  apiKeyPrefix:
                    type: string
                  active:
                    type: boolean
                  createdAt:
                    type: string
                    format: date-time
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        externalUserId:
                          type: string
                        internalUserId:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...